Fix the get action to support HTTP remote sources.
As of commit dd072c996186d691dae951911c9b07593aae29e7 the get action fails to retrieve remote sources. This is due to the File.binread method operating on File which only works with locally mounted file systems. Prior to this commit the get action used Open-URI to access the source which allowed it to be either local to the system or a remote http file.

This commit restores the use of Open-URI and also aims to maintain the binary mode reading support of the aforementioned commit.

Signed-off-by: José Valim <jose.valim@gmail.com>