Added file methods to comment and uncomment lines
A lot of times you don't want to have to replace lines.  Sometimes you want to
leave them there for future reference.

Other times you have lines (say in a common Gemfile or generated config) that
you know are there and are simply commented.

This is the solution to that problem.