Add Thor#hide_if
hide_if accepts a block, which it stores in the next-defined task, which will evaluate it whenever #hidden? is called, allowing tasks which hide and show themselves dynamically depending on arbitrary conditions.  For example, a cleanup script might only appear in the command list if there were any files that needed cleaning up.

Since Tasks can now have varied and dynamic values for #hidden, the class HiddenTask was removed as unnecessary.