Add support for background color to say
When calling set_color or say, it is now possible
to pass a background color. Additionally, the API
no longer cares about the order of color
arguments, so you can use:

  set_color(string, :red, :on_white, :bold)

When calling `say`, you can now pass an Array of
colors, which will be passed to set_color:

  say("HI!", [:red, :on_white, :bold]

This change is backwards compatible, The old
syntax (`say "HI", :red, true`) will continue to
work.