Allow calling private methods on template names
This is an expected feature from Thor. When rendering files,
Thor behaves like Sinatra. It works as if the file was embedded
in the current Thor class. Therefore, restricting the filename
to be just public methods is a wrong limitation that does not
fit with how Thor is meant to be used.