Add disable_required_check! and make disable the check in help
In Thor 0.19.2 it was added a new class method disable_class_options.
It was added to disable the check for required options in the help
command. This new method disabled the option parser for help, so it is
not possible to pass option to that command anymore.

That change introduced a regression and it is fixing the problem at the
wrong level. What we want is to disable the required check not the
option parsing.

So, we reverted the disable_class_options change and added a new method
disable_required_check! to disable the required check in the commands we
want.

Theoretically this is a breaking change but the disable_class_options
introduction was also a breaking change so it is better to revert it and
fix the problem at the correct place.

Related with #462 and #363.