Added some new features
Added the “bright” colors to the ANSI color constants
Added the ability to specify a range in Shell#ask — This doesn’t convert to an array for the answer set, it just displays the min and max of the array.
Added a `ask_if_nil` function to ask a question if the provided value is nil, otherwise return the value. — This helps with values that are required but may be nil at initialization of the task.