Do not require did_you_mean
This caused a drastic performance hit (100x slower) on our JRuby applications because it was requiring `did_you_mean` even though we have:

```
JRUBY_OPTS="--disable:did_you_mean"
```

set. In looking through other libraries that use `did_you_mean`, it looks like a fairly common pattern is to just check if the portions of `did_you_mean` that are needed are defined. I tried to replicate that in this patch.