Adds a top level package description
There could be cases where users have a very short name for the package
name, which is the cli name, in such cases it would be good to provide a
way to add a top-level package description just like there is for
sub-commands.

Example.

With the following class:
```
class Marvel < Thor
  package_name "marvel"
  package_description "A command-line interpreter for the Marvel Corporation."
end
```

We would see the follosing help:
```
$ marvel help
A command-line interpreter for the Marvel Corporation.

marvel commands:
```

Signed-off-by: Salim Afiune <afiune@chef.io>