Amend previous --skip-foo with trailing option fix
The fix in bdb43eb unintentioanlly dropped support for specifying things
like `--skip-foo=f` or `--skip-foo true`. In the first case, for no good
reason, in the second, because "true" is now interpreted as a command
option, not as a flag.

This commit restores support for those, so that arguments right after
`--skip-foo` flags will never be interpreted as the flag's value, unless
they are "true", "t", "false", or "f".

So

`my_program --skip-foo true`

means

"run my_program without arguments and the 'foo' flag unset".

whereas

`my_program --skip-foo asdf`

means

"run my_program with argument 'asdf' and the 'foo' flag unset"