Fix TypeError when no config given and warn if file unchanged
If no :before nor :after parameter was supplied, Thor now decides
to append the data after the last piece of information within the file.

In addition, a constant Hash with warnings has been added.
If the file is unchanged because the flag value couldn't be found,
Thor now warns the users and briefly explains what happened.