Fixes a bug in the calculation of the print_wrapped method
Fixes a bug that was causing the print_wrapped method to incorrectly
wrap the text and print beyond the edge of the terminal.

The reason for the bug is that the function assumed that the iteration
started with the first word but it actually started with the second word
and the first word was already in memo.