Respect the updated NO_COLOR specification
Since implementation the NO_COLOR specification has changed. Previously
if NO_COLOR was set in the environment it should be respected regardless
of value. In the new updated specification NO_COLOR should only be
respected if set to any non-empty string.

See this comment thread on the original commit for more information.
https://github.com/rails/thor/commit/21e65684432c0fcc975c1d6fda497da24149f763#r80414525

Fixes #795